+++ lib/td_cache/nonce.ex
@@ -1,0 +1,12 @@
+defmodule TdCache.Nonce do
+  @moduledoc """
+  Provides a way of generating random nonces.
+  """
+
+  def new(length) do
+    length
+    |> :crypto.strong_rand_bytes()
+    |> Base.url_encode64()
+    |> binary_part(0, length)
+  end
+end