+++ lib/td_cache/templates/preprocessor.ex
@@ -11,9 +11,10 @@ defmodule TdCache.Templates.Preprocessor do
     user_roles = AclLoader.get_roles_and_users(domain_ids)
     user_group_roles = AclLoader.get_roles_and_groups(domain_ids)
 
-    context = context
-    |> Map.put(:user_roles, user_roles)
-    |> Map.put(:user_group_roles, user_group_roles)
+    context =
+      context
+      |> Map.put(:user_roles, user_roles)
+      |> Map.put(:user_group_roles, user_group_roles)
 
     preprocess_template_content(template, context)
   end