+++ test/test_helper.exs
@@ -1,1 +1,2 @@
+TdCache.Redix.del!()
 ExUnit.start()