+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 ### Fixed
 
 - [TD-5011] `TemplateCache.put/2` was leaving stale keys in
-  `templates:name_to_id` hash
+  `templates:name_to_id` hash and `TemplateCache.list/0` was returning
+  duplicates
 
 ## [4.47.0] 2022-06-27